Server code gets own folder, first typegoose work
This was done to deal with a typescript/vscode error. vscode does not
look at any tsconfig that is not tsconfig.json and there is not setting
you can use to change this.  This ment that VSCode would give incorrect
typescript error while editing server files because tsconfig.server.json
would not be used for linting.  When I tryed out typegoose this became
to much and I had to fix it.  Typegoose add for the Device model.